+++ .changeset/refactor-cli-utils.md
@@ -1,0 +1,5 @@
+---
+"task-master-ai": patch
+---
+
+Refactor CLI utility functions into utils-cli.js
